Monopoly_Railroads
One of the railroad properties in the original Monopoly is not an actual railroad. Write a program to display the names of the four Monopoly railroad properties and ask the user to identify the property that is not a railroad. The user should be informed if the selection is correct or not. Your program should give the user an option to continue guessing or exit the program after each incorrect guess.