Create Factorial Calculator
With a user input, enter a positive number and calculate its factorial using a main() function to control the program flow. Write a function to get an integer from the user and verify the user entered a positive integer value. Write a different function to calculate the factorial for a number.