Create Raise Calculator
Write a program that asks an employee to enter their current annual salary, first name, and last name. Calculate the employee's salary for next year. Employees who earn less than $50,000 will receive a 5% raise, while those who earn $50,000 or more will receive a raise of $2,500 plus 2% of the amount over $50,000. Use the main module to control the program flow. Write a function to display the program overview, collect the inputs, calculate the raise, and display the old salary, the amount of the raise, the effective raise rate (i.e., raise / old salary), and the new salary for the employee.