Create Rose Bowl Wins
The Rosebowl.txt data set contains the names of the teams that won the Rose Bowl from 1902 (Michigan on the first line) to 2020 (Oregon on the last line). Write a program that reads the file and counts the number of Rose Bowl wins for each team (university). Create a new comma-delimited file that includes the team and the number of times the team won the Rose Bowl for each team. Display a list to the user of all teams that have won more than four Rose Bowls.