Create Alphabetical Order
Write a program that asks the user to enter a word. Then determine if the letters of the word appear in alphabetical order. Inform the user if the letters are in alphabetical order or not. Some examples of words whose letters are in alphabetical order are biopsy, adept, knotty, and affix.